Factor network files into class-per-file model.
