+++ include/bitcoin/bitcoin/network/hosts.hpp
@@ -20,6 +20,8 @@
 #ifndef LIBBITCOIN_HOSTS_HPP
 #define LIBBITCOIN_HOSTS_HPP
 
+#include <cstddef>
+#include <cstdint>
 #include <string>
 #include <functional>
 #include <system_error>
@@ -32,7 +34,7 @@
 namespace libbitcoin {
 namespace network {
 
-class hosts
+class BC_API hosts
 {
 public:
     typedef std::function<void (const std::error_code&)> load_handler;
@@ -47,27 +49,26 @@ public:
     typedef std::function<void (const std::error_code&, size_t)>
         fetch_count_handler;
 
-    BC_API hosts(threadpool& pool, const std::string& path, 
-        size_t capacity=1000);
-    BC_API ~hosts();
+    hosts(threadpool& pool, const std::string& path, size_t capacity=1000);
+    ~hosts();
 
     hosts(const hosts&) = delete;
     void operator=(const hosts&) = delete;
 
-    BC_API void load(load_handler handle_load);
-    BC_API void save(save_handler handle_save);
+    void load(load_handler handle_load);
+    void save(save_handler handle_save);
 
-    BC_API void store(const network_address_type& address,
+    void store(const network_address_type& address,
         store_handler handle_store);
-    BC_API void remove(const network_address_type& address,
+    void remove(const network_address_type& address,
         remove_handler handle_remove);
-    BC_API void fetch_address(fetch_address_handler handle_fetch);
-    BC_API void fetch_count(fetch_count_handler handle_fetch);
+    void fetch_address(fetch_address_handler handle_fetch);
+    void fetch_count(fetch_count_handler handle_fetch);
 
     /// Deprecated, set hosts path on construct.
-    BC_API hosts(threadpool& pool, size_t capacity=1000);
-    BC_API void load(const std::string& path, load_handler handle_load);
-    BC_API void save(const std::string& path, save_handler handle_save);
+    hosts(threadpool& pool, size_t capacity=1000);
+    void load(const std::string& path, load_handler handle_load);
+    void save(const std::string& path, save_handler handle_save);
 
 private:
     struct hosts_field
