+++ src/network/channel_stream_loader.cpp
@@ -1,4 +1,4 @@
-/*
+/**
  * Copyright (c) 2011-2013 libbitcoin developers (see AUTHORS)
  *
  * This file is part of libbitcoin.
@@ -17,49 +17,35 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef LIBBITCOIN_SHARED_CONST_BUFFER_HPP
-#define LIBBITCOIN_SHARED_CONST_BUFFER_HPP
+#include <bitcoin/bitcoin/network/channel_stream_loader.hpp>
 
-#include <memory>
-#include <boost/asio.hpp>
+#include <string>
+#include <bitcoin/bitcoin/network/channel_loader_module.hpp>
+#include <bitcoin/bitcoin/utility/assert.hpp>
 #include <bitcoin/bitcoin/utility/data.hpp>
 
 namespace libbitcoin {
 namespace network {
 
-// A reference-counted non-modifiable buffer class.
-class BC_API shared_const_buffer
+channel_stream_loader::~channel_stream_loader()
 {
-public:
-     // Construct from a stream object.
-     explicit shared_const_buffer(const data_chunk& user_data)
-     : data_(std::make_shared<data_chunk>(
-            std::begin(user_data), std::end(user_data))),
-        buffer_(boost::asio::buffer(*data_))
-    {
-    }
+    for (auto module: modules_)
+        delete module;
+}
 
-    // Implement the ConstBufferSequence requirements.
-    typedef boost::asio::const_buffer value_type;
-    typedef const value_type* const_iterator;
+void channel_stream_loader::add(channel_loader_module_base* module)
+{
+    BITCOIN_ASSERT(module != nullptr);
+    modules_.push_back(module);
+}
 
-    const_iterator begin() const
-    {
-        return &buffer_;
-    }
-
-    const_iterator end() const
-    {
-        return &buffer_ + 1;
-    }
-
-private:
-    std::shared_ptr<data_chunk> data_;
-    value_type buffer_;
-};
+void channel_stream_loader::load_lookup(const std::string& symbol,
+    const data_chunk& stream) const
+{
+    for (auto module: modules_)
+        if (module->lookup_symbol() == symbol)
+            module->attempt_load(stream);
+}
 
 } // namespace network
 } // namespace libbitcoin
-
-#endif
-
